add fetch service for PUT
