+++ src/Services/privateApiService.js
@@ -1,16 +1,21 @@
 import axios from "axios";
 
-const config = {
-  headers: {
-    Group: 01, //Aqui va el ID del equipo!!
-  },
+const baseUrl = "http://ongapi.alkemy.org/api";
+
+const Put = async (endPoint, id, body) => {
+  const url = id ? `${baseUrl}/${endPoint}/${id}` : `${baseUrl}/${endPoint}`;
+
+  try {
+    const response = await axios.put(url, body, {
+      headers: {
+        Authorization: `Bearer ${localStorage.getItem("token")}`,
+      },
+    });
+
+    return response.data;
+  } catch (error) {
+    return { success: false };
+  }
 };
 
-const Get = () => {
-  axios
-    .get("https://jsonplaceholder.typicode.com/users", config)
-    .then((res) => console.log(res))
-    .catch((err) => console.log(err));
-};
-
-export default Get;
+export { Put };
